auth_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg[binary]==3.2.3
bcrypt==4.2.1
PyJWT==2.10.1
